Validation DELETE and GET comment system

+++ src/controllers/strayAnimalController.js
@@ -404,69 +404,6 @@ function isExternalUrl(url) {
 }
 
 // ----------------- Edit status adoption request for a stray animal by ID -------------------------------------------
-// const updateAdoptionRequestStatus = async (req, res) => {
-//   try {
-//     // Call getLoggedInUserDataNoRes to retrieve logged-in user's data
-//     const loggedInUser = await loggedInUserService.getLoggedInUserDataNoRes(req)
-
-//     const loggedInUserRole = loggedInUser.role
-//     const loggedInUserId = loggedInUser._id.toString()
-
-//     // Call the validation function to ensure owner/admin authorization
-//     validate(req, res, async () => {
-//       // Retrieve the existing adoption request
-//       const existingAdoptionRequest = await AdoptionRequest.findById(req.params.reqId)
-
-//       if (!existingAdoptionRequest) {
-//         console.log('Adoption request not found')
-//         console.log('---------------------------------------------')
-//         return res.status(404).json({ message: 'Adoption request not found' })
-//       }
-
-//       // Check if the authenticated user is the owner of the stray animal or an admin
-//       const strayAnimal = existingAdoptionRequest.strayAnimal;
-//       if (!strayAnimal || !strayAnimal.owner || !strayAnimal.owner.ownerId) {
-//         console.log('Invalid adoption request: Missing owner information');
-//         console.log('---------------------------------------------');
-//         return res.status(400).json({ message: 'Invalid adoption request: Missing owner information' });
-//       }
-
-//       const ownerId = strayAnimal.owner.ownerId.toString();
-//       if (loggedInUserRole !== 'admin' && ownerId !== loggedInUserId) {
-//         console.log('You are not authorized to edit this adoption request');
-//         console.log('---------------------------------------------');
-//         return res.status(403).json({ message: 'You are not authorized to edit this adoption request' });
-//       }
-
-//       // Validate and update the adoption request status
-//       const { status } = req.body
-//       if (!status || (status !== 'On Request' && status !== 'Accepted')) {
-//         console.log('Invalid status provided')
-//         console.log('---------------------------------------------')
-//         return res.status(400).json({ message: 'Invalid status provided' })
-//       }
-
-//       existingAdoptionRequest.status = status
-//       existingAdoptionRequest.updatedOn = new Date()
-//       await existingAdoptionRequest.save()
-
-//       // Respond with only required fields
-//       const updateStatus = {
-//         status: existingAdoptionRequest.status,
-//         _id: existingAdoptionRequest._id,
-//         updatedOn: existingAdoptionRequest.updatedOn
-//       }
-
-//       res.json({ message: 'Updated adoption request status:', updateStatus })
-//       console.log('Updated adoption request status:', updateStatus )
-//       console.log('---------------------------------------------')
-//       console.log('Updated adoption request status successfully:', existingAdoptionRequest )
-//     })
-//   } catch (error) {
-//     console.error('Error updating adoption request:', error)
-//     res.status(500).json({ message: 'Error updating adoption request' })
-//   }
-// }
 const updateAdoptionRequestStatus = async (req, res) => {
   try {
     // Retrieve the logged-in user's data
@@ -493,12 +430,12 @@ const updateAdoptionRequestStatus = async (req, res) => {
     adoptionRequest.updatedOn = new Date()
     await adoptionRequest.save()
 
-          //  Respond with only required fields
+    //  Respond with only required fields
     const updateStatus = {
-      status: adoptionRequest.status,
-      _id: adoptionRequest._id,
-      updatedOn: adoptionRequest.updatedOn
-      }
+    status: adoptionRequest.status,
+     _id: adoptionRequest._id,
+    updatedOn: adoptionRequest.updatedOn
+    }
 
     res.json({ message: 'Updated adoption request status:', updateStatus })
     console.log('Updated adoption request status successfully by:', loggedInUser.username)
@@ -683,7 +620,7 @@ const createComment = async (req, res) => {
     )
     console.log('User', addComment.user)
     console.log('---------------------------------------------')
-    console.log('Comment: ', addComment.text)
+    console.log('Comment: ', addComment.comment)
     console.log('---------------------------------------------')
   })
   } catch (error) {
@@ -695,21 +632,19 @@ const createComment = async (req, res) => {
 
 // ----------------- Get comments for a post --------------------------
 const getComments = async (req, res) => {
-  // console.log('Fetching comments for a post')
-  // try {
-  //   const { saId } = req.params
-  //   const comments = await Comment.find({ saId })
+
   try {
     const { saId } = req.params
     const comments = await Comment.find({ 'post.saId': saId }).sort({ createdOn: -1 })
-    if (!comments) {
-      console.log('Stray animal not founnd')
+
+    if (!comments || comments.length === 0) {
+      console.log('No comments found for the specified post')
       console.log('---------------------------------------------')
-      return res.status(404).json({ message: 'Stray animal not found' })
+      return res.status(404).json({ message: 'No comments found for the specified post' })
     }
     
     res.json(comments)
-    console.log('Comments fetched successfully', comments);
+    console.log('Comments fetched successfully', comments)
     console.log('---------------------------------------------')
   } catch (error) {
     console.error('Can not get this comment :', error.message)
@@ -755,19 +690,33 @@ const deleteComment = async (req, res) => {
   console.log('Deleting a comment')
   try {
     const { commentId } = req.params
-    await Comment.findByIdAndDelete(commentId)
+    const loggedInUserId = req.user.userId 
 
-    if (!commentId) {
-      console.log('Stray animal not founnd')
+    // Find the comment by ID
+    const comment = await Comment.findById(commentId)
+
+    // Check if the comment exists
+    if (!comment) {
+      console.log('Comment not found')
       console.log('---------------------------------------------')
-      return res.status(404).json({ message: 'Stray animal not found' })
+      return res.status(404).json({ message: 'Comment not found' })
     }
+
+    // Check if the logged-in user is the owner of the comment
+    if (comment.user.userId !== loggedInUserId) {
+      console.log('Unauthorized: Only the owner can delete the comment')
+      console.log('---------------------------------------------')
+      return res.status(403).json({ message: 'Unauthorized: Only the owner can delete the comment' })
+    }
+
+    // Delete the comment
+    await Comment.findByIdAndDelete(commentId)
 
     console.log('Comment deleted successfully')
     res.json({ message: 'Comment deleted successfully' })
     console.log('---------------------------------------------')
   } catch (error) {
-    console.error('Error deleting comment:', error.message);
+    console.error('Error deleting comment:', error.message)
     console.log('---------------------------------------------')
     res.status(500).json({ message: 'Unable to delete comment' })
   }
